+++ qiskit/optimization/algorithms/__init__.py
@@ -2,7 +2,7 @@
 
 # This code is part of Qiskit.
 #
-# (C) Copyright IBM 2019, 2020.
+# (C) Copyright IBM 2020.
 #
 # This code is licensed under the Apache License, Version 2.0. You may
 # obtain a copy of this license in the LICENSE.txt file in the root directory
@@ -11,26 +11,23 @@
 # Any modifications or derivative works of this code must retain this
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
+
 """
+Optimization Algorithms (:mod:`qiskit.optimization.algorithms`)
 ====================================================================
-Optimization application stack for Aqua (:mod:`qiskit.optimization`)
-====================================================================
-This is the finance domain logic....
+Algorithms for optimization problems
 
-.. currentmodule:: qiskit.optimization
+.. currentmodule:: qiskit.optimization.algorithms
 
-Submodules
-==========
+Algorithms
+=============
 
 .. autosummary::
-   :toctree:
+   :toctree: ../stubs/
+   :nosignatures:
 
-   ising
+   grover_minimum_finder
 
 """
 
-from ._logging import (get_qiskit_optimization_logging,
-                       set_qiskit_optimization_logging)
-
-__all__ = ['get_qiskit_optimization_logging',
-           'set_qiskit_optimization_logging']
+from .grover_minimum_finder import GroverMinimumFinder