+++ qiskit/optimization/results/grover_optimization_results.py
@@ -15,7 +15,8 @@
 
 class GroverOptimizationResults:
 
-    def __init__(self, optimum_input, optimum_output, operation_counts, rotations, n_input_qubits, n_output_qubits, f):
+    def __init__(self, optimum_input, optimum_output, operation_counts, rotations, n_input_qubits,
+                 n_output_qubits, f):
         """ Stores the result of a Grover optimization problem.
             :param optimum_input: (int) The input that corresponds to the optimum output.
             :param optimum_output: (int) The optimum output value.
@@ -23,8 +24,8 @@ class GroverOptimizationResults:
             :param rotations: The total number of Grover rotations performed.
             :param n_input_qubits: The number of qubits used to represent the input.
             :param n_output_qubits: The number of qubits used to represent the output.
-            :param f: A dictionary representation of the function, where the keys correspond to a variable, and the
-                values are the corresponding coefficients.
+            :param f: A dictionary representation of the function, where the keys correspond to a
+                variable, and the values are the corresponding coefficients.
         """
         self._optimum_input = optimum_input
         self._optimum_output = optimum_output